requirements.txt
requests
beautifulsoup4
pandas
lxml
google-generativeai
selenium
openpyxl 
webdriver-manager